+++ greenfield/eldenring/data.py
@@ -5256,4 +5256,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:5b6981dec2eae1b42f6948727403d2670fb15ac37051e1575d7eeb03d3fe1daa', 'module': 'data.py', 'body_sha256': 'sha256:fef8f97a50f816557b3b04e382753ec6e415dc1453e3ea23380fa5fb120f5f32'}
+_GEN_STAMP = {'inputs_hash': 'sha256:3289f37bb95dd24d4f51279d0c1233fa5da190215ccc5755d70393dbdf504dd8', 'module': 'data.py', 'body_sha256': 'sha256:fef8f97a50f816557b3b04e382753ec6e415dc1453e3ea23380fa5fb120f5f32'}